python/requirements.txt
numpy>=1.22.4
pemja>=0.5.5,<0.5.6

anthropic==0.88.0
watchdog==6.0.0

# Redis client
redis>=5.0.0,<8.0.0
redisvl>=0.13.0

# Kafka client
confluent-kafka>=2.3.0
kafka-python>=2.0.2

# Database clients
psycopg2-binary>=2.9.0

spacy>=3.7.0,<4.0.0
sentence-transformers>=2.2.0
keybert>=0.8.0
torch>=2.0.0
transformers>=4.30.0
optimum[onnxruntime]>=1.17.0
ftfy>=6.1.0
scikit-learn>=1.0.0

trafilatura>=1.6.0
selenium>=4.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.28.0

regex>=2023.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
tqdm>=4.60.0

langchain>=0.1.0
langchain-text-splitters>=0.0.1

tabulate>=0.9.0
rich>=13.0.0

openai>=1.0.0
tiktoken>=0.5.0

apache-flink>=2.2.0

opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0

streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0